+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-04-19-1
+
+Changes:
+
+- `flux-open-id-connect-authentication-backend` authentication
+
 ## v2023-04-18-1
 
 Changes: